Поменял разрядность

+++ Task008.1/Program.cs
@@ -1,18 +1,18 @@
 ﻿//Напишите программу, которая принимает на вход число "N" и проверяет, является ли оно палиндромом.
 Console.Clear();
-int Prompt (string message)
+UInt64 Prompt (string message)
 {
 Console.Write("Введите число =>");
 String Inputed = Console.ReadLine();
-int res = Convert.ToInt32(Inputed);
+UInt64 res = Convert.ToUInt64(Inputed);
 return res;
 }
-int num = Prompt("Введите число =>");
-int left = num;
-int rev = 0;
+UInt64 num = Prompt("Введите число =>");
+UInt64 left = num;
+UInt64 rev = 0;
 while(Convert.ToBoolean(left)) 
 {
-   int r = left % 10;   
+   UInt64 r = left % 10;   
    rev = rev * 10 + r;
    left = left / 10;  
 }
